Farm out meta arguments from closure macro
The more complicated implementation has these meta args inline during
the macro call. The reason this is complicated is because forms must
be sanitized.

It's reasonable. But this implementation is much easier, at the
expense of making all calls look like:

(with-circe-actions-closure (:event-signature "event")
  (..(sexp)))

The alternative, much sleeker implementation allows this
(with-circe-actions-closure
  :event-signature "event"
  (..(sexp)))

This is possible but looking at how use-package does it, it looks
mighty complicated. Maybe I'll tackle it tonight, after we break fast.